@@ -1,9 +1,9 @@
 {
   "name": "mmf-blog-vite-react",
-  "version": "2.0.0",
-  "author": "lincenying <lincenying@qq.com>",
-  "private": true,
   "type": "module",
+  "version": "2.0.0",
+  "private": true,
+  "author": "lincenying <lincenying@qq.com>",
   "scripts": {
     "prepare": "husky install",
     "serve": "vite",
@@ -13,20 +13,20 @@
     "lint:fix": "eslint ./src --ext .js,.jsx,.ts,.tsx --fix"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.9.2",
-    "ahooks": "^3.7.4",
-    "antd": "^5.2.1",
-    "axios": "^1.3.5",
+    "@reduxjs/toolkit": "^1.9.5",
+    "ahooks": "^3.7.7",
+    "antd": "^5.5.1",
+    "axios": "^1.4.0",
     "history": "^5.3.0",
-    "lcy-utils": "^0.2.2",
+    "lcy-utils": "^0.3.0",
     "md5": "^2.3.0",
     "nprogress": "^0.2.0",
     "prop-types": "^15.8.1",
-    "qs": "^6.11.0",
+    "qs": "^6.11.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
-    "react-router-dom": "^6.8.1",
+    "react-router-dom": "^6.11.2",
     "react-router-redux": "^4.0.8",
     "react-transition-group": "^4.4.5",
     "redux": "^4.2.1",
@@ -34,25 +34,25 @@
     "toastr": "^2.1.4"
   },
   "devDependencies": {
+    "@lincy/eslint-config-react": "^4.0.2",
     "@types/md5": "^2.3.2",
     "@types/nprogress": "^0.2.0",
     "@types/qs": "^6.9.7",
-    "@types/react-dom": "^18.0.11",
+    "@types/react-dom": "^18.2.4",
     "@types/react-router-redux": "^5.0.22",
-    "@types/react-transition-group": "^4.4.5",
+    "@types/react-transition-group": "^4.4.6",
     "@types/toastr": "^2.1.40",
-    "@vitejs/plugin-react": "^3.1.0",
-    "eslint": "^8.38.0",
-    "eslint-config-lcy-react": "^4.0.1",
+    "@vitejs/plugin-react": "^4.0.0",
+    "eslint": "^8.41.0",
     "husky": "^8.0.3",
-    "lint-staged": "^13.2.1",
-    "postcss": "^8.4.21",
-    "sass": "^1.61.0",
+    "lint-staged": "^13.2.2",
+    "postcss": "^8.4.23",
+    "sass": "^1.62.1",
     "typescript": "^5.0.4",
-    "unocss": "^0.50.8",
-    "unplugin-auto-import": "^0.15.2",
-    "vite": "^4.1.1",
-    "vite-plugin-pwa": "^0.14.4",
+    "unocss": "^0.52.3",
+    "unplugin-auto-import": "^0.16.2",
+    "vite": "^4.3.8",
+    "vite-plugin-pwa": "^0.15.1",
     "vite-plugin-style-import": "^2.0.0"
   },
   "lint-staged": {
