@@ -1,13 +1,15 @@
 import React from 'react'
 import { Navigate } from 'react-router-dom'
+import { globalState } from '@/store/global'
 
 interface Props {
-    [x: string]: any
-    component: any
+    component: React.ComponentType<any>
+    location?: string
 }
 
-const FrontendAuthorized = (props: Props) => {
+function FrontendAuthorized(props: Props) {
     const { component: Component, ...rest } = props
-    return rest.global.cookies.user ? <Component {...rest} /> : <Navigate to={{ pathname: '/' }} state= {{ from: rest.location }} replace />
+    const global = useSelector(globalState)
+    return global.cookies.user ? <Component {...rest} /> : <Navigate to={{ pathname: '/' }} state= {{ from: rest.location }} replace />
 }
 export default FrontendAuthorized
