@@ -2,7 +2,7 @@ import React from 'react'
 import SignIn from '@/components/sign-in.jsx'
 import SignUp from '@/components/sign-up.jsx'
 
-const Sign = () => {
+function Sign() {
     const location = useLocation()
     const backend = useRef(location.pathname.indexOf('/backend') === 0)
     const signUpHtml = !backend.current ? <SignUp /> : ''
