@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 import api from '@/api'
 import { categoryState, deleteCategory, getCategoryList, recoverCategory } from '@/store/global/category'
 
-const CategoryList = () => {
+function CategoryList() {
     const dispatch = useDispatch()
     const category = useSelector(categoryState)
 
@@ -36,16 +36,8 @@ const CategoryList = () => {
         let btn
         if (!item.cate_num) {
             btn = item.is_delete
-                ? (
-                <a onClick={handleRecover.bind(null, item._id)} href={undefined}>
-                    恢复
-                </a>
-                    )
-                : (
-                <a onClick={handleDelete.bind(null, item._id)} href={undefined}>
-                    删除
-                </a>
-                    )
+                ? <a onClick={handleRecover.bind(null, item._id)} href={undefined}>恢复</a>
+                : <a onClick={handleDelete.bind(null, item._id)} href={undefined}>删除</a>
         }
 
         return (
@@ -53,9 +45,7 @@ const CategoryList = () => {
                 <div className={`list-title${item.is_delete ? ' text-red-500 line-through' : ''}`}>{item.cate_name}</div>
                 <div className="list-time">{item.cate_order}</div>
                 <div className="list-action">
-                    <Link to={`/backend/category/modify/${item._id}`} className="badge badge-success">
-                        编辑
-                    </Link>
+                    <Link to={`/backend/category/modify/${item._id}`} className="badge badge-success">编辑</Link>
                     {btn}
                 </div>
             </div>
