@@ -11,8 +11,9 @@ import { articleState, getArticleItem } from '@/store/frontend/article'
 import { categoryState, getCategoryList } from '@/store/global/category'
 import { getTrending, trendingState } from '@/store/frontend/trending'
 
-const addTarget = (content: string) => {
-    if (!content) return ''
+function addTarget(content: string) {
+    if (!content)
+        return ''
     return content.replace(/<a(.*?)href=/g, '<a$1target="_blank" href=')
 }
 
@@ -30,9 +31,12 @@ export default function Article() {
     useMount(async () => {
         console.log('article useMount: start')
         window.scrollTo(0, 0)
-        if (article.pathname !== location.pathname) dispatch(await getArticleItem({ id, pathname }))
-        if (category.lists.length === 0) dispatch(await getCategoryList())
-        if (trending.data.length === 0) dispatch(await getTrending())
+        if (article.pathname !== location.pathname)
+            dispatch(await getArticleItem({ id, pathname }))
+        if (category.lists.length === 0)
+            dispatch(await getCategoryList())
+        if (trending.data.length === 0)
+            dispatch(await getTrending())
         console.log('article useMount: end')
     })
 
