@@ -53,11 +53,13 @@ const reducers = {
     },
     deletes: (state: UserStore, action: PayloadAction<string>) => {
         const obj = state.lists.data.find(ii => ii._id === action.payload)
-        if (obj) obj.is_delete = 1
+        if (obj)
+            obj.is_delete = 1
     },
     recover: (state: UserStore, action: PayloadAction<string>) => {
         const obj = state.lists.data.find(ii => ii._id === action.payload)
-        if (obj) obj.is_delete = 0
+        if (obj)
+            obj.is_delete = 0
     },
 }
 
@@ -75,15 +77,15 @@ export const {
     recover: recoverBackendUser,
 } = slice.actions
 
-export const getUserList = async (config: Record<string, any>) => {
-    const { code, data } = await api.get('backend/user/list', config)
+export async function getUserList(config: Record<string, any>) {
+    const { code, data } = await api.get<ResponseDataLists<User[]>>('backend/user/list', config)
     if (code === 200)
         return receiveBackendUser({ ...data, ...config })
 
     return setMessage(errConfig)
 }
-export const getUserItem = async (config: Record<string, any>) => {
-    const { code, data } = await api.get('backend/user/item', config)
+export async function getUserItem(config: Record<string, any>) {
+    const { code, data } = await api.get<User>('backend/user/item', config)
     if (code === 200)
         return insertBackendUser({ data, ...config })
 
