@@ -36,11 +36,13 @@ export const slice = createSlice({
         },
         deleteCategory: (state, action: PayloadAction<string>) => {
             const obj = state.lists.find(ii => ii._id === action.payload)
-            if (obj) obj.is_delete = 1
+            if (obj)
+                obj.is_delete = 1
         },
         recoverCategory: (state, action: PayloadAction<string>) => {
             const obj = state.lists.find(ii => ii._id === action.payload)
-            if (obj) obj.is_delete = 0
+            if (obj)
+                obj.is_delete = 0
         },
     },
 })
@@ -54,15 +56,15 @@ export const {
     recoverCategory,
 } = slice.actions
 
-export const getCategoryList = async (config?: Record<string, any>) => {
-    const { code, data } = await api.get('backend/category/list', config)
+export async function getCategoryList(config?: Record<string, any>) {
+    const { code, data } = await api.get<ResponseDataList<Category[]>>('backend/category/list', config)
     if (code === 200)
         return receiveCategoryList(data.list)
 
     return setMessage(errConfig)
 }
-export const getCategoryItem = async (config: Record<string, any>) => {
-    const { code, data } = await api.get('backend/category/item', config)
+export async function getCategoryItem(config: Record<string, any>) {
+    const { code, data } = await api.get<Category>('backend/category/item', config)
     if (code === 200)
         return receiveCategoryItem(data)
 
